Added i386 testhat tests
I created a function to detect architecture and compare the appropriate
files. However, now I can't run Check within RStudio in Windows, because
it doesn't skip the parallel tests and results in an error. The tests
alone run fine, but the Check only succeeds if done in a cmd window.
Linux should be fine.